requirements.txt

RPi.GPIO
spidev
RTIMU
paho-mqtt
hx711v0_5_1
numpy
matplotlib
